[docs] comments on evenFlood and globalEvenFlood
There's a very real disconnect where I understand that this solution, 
globalEvenFlood, will work for all cases (maybe) and how to explain it.

I "feel" like this is the solution but it may not be unless I can 
clearly articulate what how exactly parts of the solution contribute to 
the exact solution.

Look into writing proofs for logical flow like this.