[bug] undefined being passed when there were 0 length arrays
worked with greater-than but didn't work with not-lesser-than or equal, 
but equal is very rarely hit, so it never showed up.

reason: since undefined was returned for the array access of a 0 length 
array, a comparison would always return false (even undefined == 
undefined). So the greater-than condition would always work since a 0 
length array can never be greater than an array with some length, but a 
not-lesser-than condition would never work since a 0 length array would 
always be lesser but undefined would make it return false. Same goes for 
equals, both rows could be 0 length, but return false anyways due to 
undefined.

Actually in the case of evenFlood and globalEvenFlood are never two 0 
length arrays being compared so it's moot but good to know.

also fixed bug with calculating longestIndex