[docs] shoulda really keyed in on this when I implemented them
balancing moral and functionality is more difficult when I can't put 
something concrete and proactive on the page

idea: when coming up with an algorithm you think will work but are 
unsure, just add a "splat" prefix or something so you can minimize 
documentation and represent it accurately. Like a brainstorming bit of 
code, scrap paper basically.